more rename and reorg
