+++ frame/support/src/lib.rs
@@ -64,9 +64,9 @@ pub mod event;
 pub mod inherent;
 #[macro_use]
 pub mod error;
-pub mod base_origin;
 pub mod instances;
 pub mod migrations;
+pub mod origin;
 pub mod traits;
 pub mod weights;
